added authentication for secure routes

+++ src/jobs/jobs-service.js
@@ -1,6 +1,6 @@
 const jobsService = {
-   getAllJobs(knex) {
-       return knex.select('*').from('jobs');
+   getUsersJobs(knex, user_id) {
+       return knex.select('*').from('jobs').where({user_id});
    },
    insertJob(knex, newJob) {
        return knex
@@ -10,14 +10,14 @@ const jobsService = {
        .then(rows =>  rows[0])
 
    },
-   getJobById(knex, id) {
-    return knex.from('jobs').select('*').where({ id }).first()
+   getJobById(knex, id, user_id) {
+    return knex.from('jobs').select('*').where({ id, user_id }).first()
      },
-   deleteJob(knex, id) {
-       return knex('jobs').where({ id }).delete()
+   deleteJob(knex, id, user_id) {
+       return knex('jobs').where({ id,user_id }).delete()
    },
-   updateJob(knex, id, updateJob) {
-       return knex('jobs').where({ id }).update(updateJob)
+   updateJob(knex, id, updateJob, user_id) {
+       return knex('jobs').where({ id,user_id }).update(updateJob)
    }
 }
